requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.10
